Updated the model undeployment REST API endpoint
The HTTP status code of a successful delete operation was changed from
204 (No Content) to 200 (OK). The response body is an empty JSON object,
which is represented by class org.openscoring.common.SimpleResponse.

This change was necessitated by technical difficulties that surround
the handling of HTTP 204 status code in client applications. The case in
point is the FormPanel widget of Google Web Toolkit (GWT), which does
not fire FormPanel$SubmitCompleteEvent.