Improved the handling of evaluation errors for multi-record evaluation requests
The process is controlled by record isolation level, which may be
specified either as "record" or "all-records-or-nothing". The choice
between the two is dictated by the output data format.
Namely, "record" requires structured data format such as JSON, because
it will be necessary to distinguish between successful and failed
records, whereas "all-records-or-nothing" does not have such
requirements.